Convert JS files to TS and fix circular dependencies
