+++ package.json
@@ -3,6 +3,14 @@
   "version": "1.5.0",
   "description": "ESLint plugin suggesting potential corrections when accessing non-existent object members in TypeScript.",
   "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "exports": {
+    ".": {
+      "import": "./dist/index.js",
+      "require": "./dist/index.js"
+    }
+  },
+  "type": "commonjs",
   "files": [
     "dist/",
     "README.md"
@@ -13,15 +21,15 @@
     "lint": "eslint --ext .ts .",
     "lint:default": "eslint --ext .ts .",
     "example:build": "tsc -p example/tsconfig.json",
-    "example:lint": "eslint --ext .ts example/src/ -c example/eslint.config.js",
+    "example:lint": "eslint --ext .ts example/src/ -c example/eslint.config.ts",
     "start": "npm run build && npm run example:lint",
-    "test:typos": "eslint --ext .ts example/src/example1-typos.ts -c example/eslint.config.js",
-    "test:similarity": "eslint --ext .ts example/src/example2-similarity.ts -c example/eslint.config.js",
-    "test:edge-cases": "eslint --ext .ts example/src/example3-edge-cases.ts -c example/eslint.config.js",
-    "test:jaro-winkler": "eslint --ext .ts example/src/example4-jaro-winkler.ts -c example/eslint.config.js",
-    "test:corner-cases": "eslint --ext .ts example/src/example5-corner-cases.ts -c example/eslint.config.js",
+    "test:typos": "eslint --ext .ts example/src/example1-typos.ts -c example/eslint.config.ts",
+    "test:similarity": "eslint --ext .ts example/src/example2-similarity.ts -c example/eslint.config.ts",
+    "test:edge-cases": "eslint --ext .ts example/src/example3-edge-cases.ts -c example/eslint.config.ts",
+    "test:jaro-winkler": "eslint --ext .ts example/src/example4-jaro-winkler.ts -c example/eslint.config.ts",
+    "test:corner-cases": "eslint --ext .ts example/src/example5-corner-cases.ts -c example/eslint.config.ts",
     "test": "npm run build && npm run test:typos && npm run test:similarity && npm run test:edge-cases && npm run test:jaro-winkler && npm run test:corner-cases",
-    "lint:formatter": "eslint --ext .ts example/src/ -c example/eslint.config.js"
+    "lint:formatter": "eslint --ext .ts example/src/ -c example/eslint.config.ts"
   },
   "repository": {
     "type": "git",
@@ -38,7 +46,6 @@
   ],
   "author": "TON AI",
   "license": "MIT",
-  "type": "commonjs",
   "bugs": {
     "url": "https://github.com/ton-ai-core/eslint-plugin-suggest-members/issues"
   },
@@ -58,7 +65,9 @@
     "eslint": "^9.24.0",
     "eslint-formatter-codeframe": "^7.32.1",
     "jest": "^29.7.0",
+    "jiti": "^2.4.2",
     "ts-jest": "^29.3.1",
+    "ts-node": "^10.9.2",
     "typescript": "^5.8.3"
   },
   "publishConfig": {
