+++ tsconfig.json
@@ -3,15 +3,13 @@
     "target": "es2020",
     "module": "Node16",
     "moduleResolution": "node16",
-    "strict": true,
     "esModuleInterop": true,
+    "strict": true,
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true,
-    "outDir": "dist",
-    "declaration": true,
-    "resolveJsonModule": true,
-    "isolatedModules": true
+    "outDir": "./dist",
+    "allowSyntheticDefaultImports": true
   },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules", "dist", "tests"]
+  "include": ["src/**/*.ts", "*.ts"],
+  "exclude": ["node_modules", "dist", "example"]
 }
